Добавлен класс треугольника

+++ Lab7/Lab7/Common/Point.h
@@ -2,6 +2,6 @@
 
 struct Point
 {
-	int x;
-	int y;
+	float x;
+	float y;
 };
